Add get_sections, extract_sections, _get_hidden_resources
This is based on the get_downloadable_content method from dgorissen.
The test are adapted from the test for parse_syllabus in coursera-dl.
The most important differences are:
- sections instead of weeklyTopics
- sections, lectures and resources are dictionaries
- section, lecture, resource names are not trimmed or sanitized. This
way we can use filters on the complete name. Sanitizing will happen
when writing to filesystem
- _get_hidden_resources also extracts srt files (this replaces
grab_hidden_video_url and transform_preview_url in coursera-dl)
- files without extension are skipped, unless they are served from
cloudfront.net